Update webpack tooling
