+++ src/webpack/webpack.ts
@@ -1,104 +1,48 @@
-import * as CopyWebpackPlugin from 'copy-webpack-plugin';
+import {RequestHandler} from 'express-serve-static-core';
+import * as path from 'path';
 import * as webpack from 'webpack';
-
-import {getProjectConfig} from '../utils/getProjectConfig';
-import {getProjectPackageJson} from '../utils/getProjectPackageJson';
-
-/**
- * Custom options which can be passed into webpack.
- */
-export interface CustomWebpackOptions extends webpack.Options.Optimization {
-    /**
-     * If webpack should minify this module. Defaults to true.
-     */
-    minify?: boolean;
-    /**
-     * If the resulting module should inject itself into the window object to make itself easily accessible within HTML.  Defaults to false.
-     *
-     * Should be used in combination with the 'libraryName' option.
-     */
-    isLibrary?: boolean;
-    /**
-     * Sets the global variable name for the library on the window.  Required to have 'isLibrary' enabled.
-     */
-    libraryName?: string;
-    /**
-     * Allows a custom output file name to be used instead of the default [name]-bundle.js
-     */
-    outputFilename?: string;
-}
+import * as webpackDevMiddleware from 'webpack-dev-middleware';
+import {WebpackMode} from '../types';
+import {getRootDirectory} from '../utils/getRootDirectory';
 
 /**
- * Shared function to create a webpack config for an entry point
- */
-export function createConfig(outPath: string, entryPoint: string, options: CustomWebpackOptions, ...plugins: webpack.Plugin[]) {
-    const config: webpack.Configuration = {
-        entry: entryPoint,
-        optimization: {minimize: !options || options.minify !== false},
-        output: {path: outPath, filename: `${options && options.outputFilename || '[name]-bundle'}.js`},
-        resolve: {extensions: ['.ts', '.tsx', '.js']},
-        module: {
-            rules: [
-                {test: /\.css$/, loader: 'style-loader'},
-                {test: /\.css$/, loader: 'css-loader'},
-                {test: /\.module.css$/, loader: 'css-loader', query: {modules: true, localIdentName: '[name]__[local]___[hash:base64:5]'}},
-                {test: /\.(png|jpg|gif|otf|svg)$/, use: [{loader: 'url-loader', options: {limit: 8192}}]},
-                {test: /\.tsx?$/, loader: 'ts-loader'}
-            ]
-        },
-        plugins: []
-    };
-
-    if (options && options.isLibrary === true) {
-        if (!!options.libraryName) {
-            config.output!.library = options.libraryName;
-        } else {
-            config.output!.library = '[name]';
-        }
-        config.output!.libraryTarget = 'umd';
-    }
-    if (plugins && plugins.length) {
-        config.plugins!.push.apply(config.plugins, plugins);
-    }
-
-    return config;
-}
-
-/**
- * Provider temporarily requires an extra plugin to override index.html within provider app.json
- * Will be removed once the RVM supports relative paths within app.json files
- */
-export const manifestPlugin = (() => {
-    const {SERVICE_NAME, PORT} = getProjectConfig();
-
-    return new CopyWebpackPlugin([{
-        from: './res/provider/app.json',
-        to: '.',
-        transform: (content) => {
-            const config = JSON.parse(content);
-
-            if (typeof process.env.SERVICE_VERSION !== 'undefined' && process.env.SERVICE_VERSION !== '') {
-                config.startup_app.url = `https://cdn.openfin.co/services/openfin/${SERVICE_NAME}/` + process.env.SERVICE_VERSION + '/provider.html';
-                config.startup_app.autoShow = false;
-            } else {
-                console.warn('Using \'npm run build\' (or build:dev) when running locally. Can debug without building first by running \'npm start\'.');
-                config.startup_app.url = `http://localhost:${PORT}/provider/provider.html`;
-            }
-
-            return JSON.stringify(config, null, 4);
-        }
-    }]);
-})();
-
-
-/**
- * Replaces 'PACKAGE_VERSION' constant in source files with the current version of the service,
- * taken from the 'package.json' file.
+ * Executes Webpack.  Doubles as express-compatible middleware function to serve webpack modules.
  *
- * This embeds the package version into the source file as a string constant.
+ * Wrapper will immediately terminate the server if the initial build fails.
+ *
+ * This is a wrapper around the webpack-dev-middleware utility.
  */
-export const versionPlugin = (() => {
-    const PACKAGE_VERSION = getProjectPackageJson().version;
+export async function executeWebpack(mode: WebpackMode, writeToDisk: boolean): Promise<RequestHandler> {
+    return new Promise<RequestHandler>((resolve) => {
+        // Load config and set development mode
+        const config: webpack.Configuration[] = require(getRootDirectory() + '/webpack.config.js');
 
-    return new webpack.DefinePlugin({PACKAGE_VERSION: `'${PACKAGE_VERSION}'`});
-})();
+        config.forEach((entry: webpack.Configuration) => entry.mode = (entry.mode || mode));
+
+        // Create express middleware
+        const compiler = webpack(config);
+        const middleware = webpackDevMiddleware(compiler, {publicPath: '/', writeToDisk});
+
+        // Wait until initial build has finished before starting application
+        const startTime = Date.now();
+        middleware.waitUntilValid((result) => {
+            // result is actually {stats: webpack.Stats[]}, but the type reports it as only webpack.Stats;
+            const results = result as unknown as {stats: webpack.Stats[]};
+
+            // Output build times
+            const buildTimes = results.stats.map(stat => {
+                const component = path.relative(getRootDirectory(), stat.compilation.outputOptions.path);
+                return `${component}: ${(stat.endTime!.valueOf() - stat.startTime!.valueOf()) / 1000}s`;
+            });
+            console.log(`\nInitial build complete after ${(Date.now() - startTime) / 1000} seconds\n    ${buildTimes.join('\n    ')}\n`);
+
+            // Check build status
+            if (results.stats.find(stats => stats.compilation.errors.length > 0)) {
+                console.error('Build failed. See output above.');
+                process.exit(1);
+            } else {
+                resolve(middleware);
+            }
+        });
+    });
+}
